+++ packages/app/src/library/Signers/LedgerSigner.ts
@@ -2,61 +2,54 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { merkleizeMetadata } from '@polkadot-api/merkleize-metadata'
-import { createV4Tx, getSignBytes } from '@polkadot-api/signers-common'
-import type { V15 } from '@polkadot-api/substrate-bindings'
-import { decAnyMetadata } from '@polkadot-api/substrate-bindings'
-import type { PolkadotSigner } from 'polkadot-api'
-import { mergeUint8 } from 'polkadot-api/utils'
-import { Ledger } from '../../contexts/LedgerHardware/static/ledger'
-import { getExtraSignedExtensions } from './util'
+import { Ledger } from 'contexts/LedgerHardware/static/ledger'
+import type { ExtraSignedExtension, SubmittableExtrinsic } from 'dedot'
+import type { PayloadOptions } from 'dedot/types'
+import type { HexString } from 'dedot/utils'
+import { hexToU8a, u8aToHex } from 'dedot/utils'
 
 export class LedgerSigner {
-  #publicKey: Uint8Array
+  constructor(
+    public from: string,
+    public extraSignedExtension: (
+      signerAddress: string,
+      payloadOptions?: PayloadOptions
+    ) => ExtraSignedExtension | undefined,
+    public tx: SubmittableExtrinsic,
+    public metadata: HexString
+  ) {}
 
-  constructor(pubKey: Uint8Array) {
-    this.#publicKey = pubKey
-  }
-
-  async getPolkadotSigner(
-    networkInfo: { decimals: number; tokenSymbol: string },
+  async sign(
+    info: {
+      decimals: number
+      tokenSymbol: string
+      specName: string
+      specVersion: number
+      ss58: number
+    },
     index: number
-  ): Promise<PolkadotSigner> {
+  ) {
     const { app } = await Ledger.initialise()
 
-    const signTx: PolkadotSigner['signTx'] = async (
-      callData,
-      signedExtensions,
-      metadata
-    ) => {
-      const merkleizer = merkleizeMetadata(metadata, networkInfo)
-      const digest = merkleizer.digest()
-      const v15 = decAnyMetadata(metadata).metadata.value as unknown as V15
-      const { extra, additionalSigned } = getExtraSignedExtensions(
-        v15,
-        digest,
-        signedExtensions,
-        true
-      )
+    const merkleizer = merkleizeMetadata(this.metadata, {
+      decimals: info.decimals,
+      tokenSymbol: info.tokenSymbol,
+    })
+    const extra = this.extraSignedExtension(this.from, {
+      metadataHash: u8aToHex(merkleizer.digest()),
+    })
 
-      const toSign = mergeUint8(callData, ...extra, ...additionalSigned)
-      const { signature } = await Ledger.signPayload(
-        app,
-        index,
-        toSign,
-        merkleizer.getProofForExtrinsicPayload(toSign)
-      )
-      return createV4Tx(v15, this.#publicKey, signature, extra, callData)
+    if (!extra) {
+      return
     }
+    await extra.init()
 
-    return {
-      publicKey: this.#publicKey,
-      signTx,
-      signBytes: getSignBytes(async (x) => {
-        const { signature } = await Ledger.signPayload(app, index, x)
-        // NOTE: the signature includes a "0x00" at the beginning, indicating a ed25519 signature.
-        // this is not needed for non-extrinsic signatures.
-        return signature.subarray(1)
-      }),
-    }
+    const toSign = extra.toRawPayload(this.tx.callHex).data
+    const proof = u8aToHex(merkleizer.getProofForExtrinsicPayload(toSign))
+
+    const signature = (
+      await Ledger.signPayload(app, index, hexToU8a(toSign), hexToU8a(proof))
+    ).signature
+    return { signature: u8aToHex(signature), data: extra.data }
   }
 }
